Modernize database_cleaner & spec_helper config.
By default, RSpec runs tests in transactions. That's usually the
fastest way to handle database cleanup between tests. But transactions
apparently isolate data in a way that interferes with specs using a
separate process such as a headless browser. That applies to our
capybara "feature" specs.

So, we use the database_cleaner gem to clean up between tests. The
gem allows us to use different strategies for our capybara specs.
To maximize speed, the typical advice is to configure DatabaseCleaner
to use the "truncation" strategy for capybara specs, and transactions
everywhere else. For example, we could flag our capybara tests with
`:js => true`, and then configure DatabaseCleaner in
`spec/support/database_cleaner.rb` to use the "truncation" strategy
just for the tests with ":js => true".

Or make sure all capybara specs are in "features", and configure
those to use "truncation".

This commit fixes two problems with our DatabaseCleaner config:

- We tried to treat javascript specs differently, but we didn't flag
  or identify them, so it wasn't effective.
- If we did treat them differently, then we would use TWO strategies.

One symptom is that exceptions during our rspec runs could use up
all available database connections.  The specs would then slow to
a crawl, and fail.

Something to consider is whether the speed of transactions is worth
maintaining the correct configuration to treat feature/capybara
specs differently. I found this StackOverflow post interesting:

  http://stackoverflow.com/questions/11419536/postgresql-truncation-
  speed/11423886

So I ran the specs locally with each strategy. "transaction" and
"deletion" were both ~ 30 seconds; "truncation" was > 60 seconds.

In this commit:

- Use the "deletion" strategy between all specs.
- Update the syntax of the DatabaseCleaner config, and remove the
  possible double strategy.
- Move RSpec config block from support/database_cleaner to spec_helper.