Change Person has_many :certs to work in Ruby 2.3.
Ruby 2.3 changes Hash to respond to `to_proc`, which interferes with
using a Hash in a Rails 3.2 `conditions` clause on an association.
This commit replaces the Hash with a lambda.  Using the lambda
preserves the benefit of the Hash version: `person.certs.build`
will build a cert with the condition predefined.  In this case,
status will still be prepopulated with "Active", as it is with the
Hash version.