Adding Availability model to system. (#136)
* Adding Availability model to system.

This is a big commit for a complex area. Certainly not finished in this
area, but major functionality is here, such as tracking vacations
acrooss events. Will also need additional tests to test when an
availability and an event overlap.

* Removed extra call to bootstrap JS

* Listing availabilities on the person show and

corrections in data model. It removes the link from availabilities to
events, which can be derived from time spans.

* Removed association from events to availabilities

And made the Availability factory better

* Added scema.rb to PR

* Further removal of events from availabilities.