Issue #395: Fix start time sorting for availabilities on Person#show
Remove the unnecessary `<time>` tag from around start times in the
availability table on the Person#show page. This was causing the sorting
functionality in the datatable to misinterpret start time values and
sort alphabetically instead of by time.

I also ran into some FK failures in the test suite, and switched the
database cleaner strategy to truncation. In my experience truncation is
a quicker and more effective approach to use, and switching it resolved
the test suite issues I was seeing.