Create inspection_tag table and associate it with taggable models
Since tags are to have a many to many, polymorphic relationship with
item_types, item_categories, and inspection_questions, I added a tagging
table to handle the matching up of tags with their owners and used
taggable to interface with different models.

I also added assocations on the tag model with the taggable models so
that we can directly from tags to item_types, e.g. @tag.item_types