[staffing-level-endpoint] Create staffing level api endpoint
Why:
We would like to make the current staffing level information available
in an api endpoint for consumption by alert indicators.

This PR:
Introduces and endpoint which will return some JSON indicting the current
staffing level. A new namespace, api, was added to the routes to
separate this concern from the rest of the app. jbuilder was added as a
gem to use as the json serializer/templater.